Make publishing of the extension a manual action
Previously we automatically published the extension when the Changeset PR got merged, but now that we also bump the CLI version through changeset that means that we have to publish the extension whenever we want to release the CLI, which is not great, especially because we want to limit the number of releases of the extension.

This PR removes the PR-closed trigger, which means that to release the extension we have to manually trigger this workflow.