Never rerun  initialization.  This _might_ be idempotent, but the code it too complex to be sure.
I'm using a syntactically + semantically really ugly way to do this just so I can limit the amount of kilocode_change markers.  Ideally, it would't be the responsibility of the consumers to check whether migrations are done; that would be be done during construction, which should return asynchronously itself!  but that would be a much larger change, so... alas.