Remove custom changelog formatting in favor of default changeset output
The changesets/action with @changesets/changelog-github already generates a
well-formatted changelog with:
- PR links
- Author attribution
- Proper categorization (patch/minor/major)

The custom Python script was reformatting this output by:
- Removing the '### Patch Changes' header
- Changing version format from '## X.X.X' to '## [vX.X.X]'

These customizations are not essential and add maintenance burden. The default
changeset changelog format is widely used and well-understood.

This removes:
- .github/scripts/overwrite_changeset_changelog.py
- 'Get version' step
- 'Update Changelog Format' step
- 'Push Changelog updates' step

The workflow is now reduced to its essential function: creating a version bump
PR with the standard changeset changelog.